SDN Orchestra logs changed and manageable from the project conf file. New parcket formats to support sequence of 16 bit long implemented.

+++ os/net/sdn-net/sdn-advertisement.h
@@ -59,7 +59,7 @@ extern struct etimer na_timer_periodic; /**< Timer for periodic NA */
 
 #if !BUILD_WITH_SDN_CONTROLLER_SERIAL
 /** \brief Reset NA sequence number */
-void sdn_na_reset_seq(uint8_t new_cycle_seq);
+void sdn_na_reset_seq(uint16_t new_cycle_seq);
 #endif
 
 /** \brief Initialize ND structures */
